+++ tests/testthat/test-update.R
@@ -8,7 +8,7 @@ test_that("updating with a length-0 index makes the weights NA", {
   index <- aggregate(epr, agg)
   expect_equal(update(agg, index)[-5], agg[-5])
 
-  expect_equal(weights(update(agg, index)),
+  expect_equal(weights(update(agg, index), ea_only = FALSE),
                list(c("1" = NA_real_, "2" = NA),
                     c("11" = NA_real_, "21" = NA, "12" = NA),
                     c("111" = NA_real_, "211" = NA, "121" = NA, "112" = NA)))
@@ -17,11 +17,11 @@ test_that("updating with a length-0 index makes the weights NA", {
 test_that("an index that doesn't line up with the pias introduces NA", {
   epr <- elemental_index(1:2, ea = c("111", "121"))
   index <- aggregate(epr, agg)
-  expect_equal(weights(update(agg, index)),
+  expect_equal(weights(update(agg, index), ea_only = FALSE),
                list(c("1" = NA_real_, "2" = NA),
                     c("11" = NA_real_, "21" = NA, "12" = 2),
                     c("111" = 1, "211" = NA, "121" = 2, "112" = NA)))
-  expect_equal(weights(update(agg, index, r = 0)),
+  expect_equal(weights(update(agg, index, r = 0), ea_only = FALSE),
                list(c("1" = NA_real_, "2" = NA),
                     c("11" = NA_real_, "21" = NA, "12" = 1),
                     c("111" = 1, "211" = NA, "121" = 1, "112" = NA)))